pyproject.toml
[tool.poetry]
name = "flylo"
version = "0.1.0"
description = ""
authors = ["Cam Crow <19ccrow99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.124.4"
uvicorn = "^0.38.0"
black = "^25.12.0"
isort = "^7.0.0"
ruff = "^0.14.9"
amadeus = "^12.0.0"
python-dotenv = "^1.2.1"
requests = "^2.32.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
